Fix digital distortion
There were several issues causing digital distortion.

The main culprit was a bug in AnalogInput which was always detecting a
change in all the parameters all the time. This resulted in the DSS
simulation getting hammered with parameter changes as the sound was
being calculated which introduced a shitton of distortion.

The second issue was that the analog frames were being processed twice
which wasn't causing any noticable issues but it's simply unnecessary.

The third issue was that we were allowing parameter changes to happen
as frequently as the user wanted, however according to the paper it
should not happen more frequently than every 20 audio frames.